Fixed bug where clearing the VerticalObjectEditor multiple times would keep calling old callback. Fixed bug where instances would delete their files, but immediately recreate them when the focus on the node in the tree was lost.
git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@545 146ae76c-2b19-ef4c-9a69-67eb2b888ad7

+++ Anomaly/Controller/ProjectManagement/InstanceGroup.cs
@@ -207,9 +207,15 @@ namespace Anomaly
 
         private void destroySimObjectCallback(EditUICallback callback, EditInterfaceCommand command)
         {
-            String instanceFile = instanceFileManager.resolveSourceObject(callback.getSelectedEditInterface());
-            InstanceWriter.Instance.deleteInstance(this, instanceFile);
-            removeInstanceFile(instanceFile);
+            EditInterface editInterface = callback.getSelectedEditInterface();
+            if (editInterface.hasEditableProperties())
+            {
+                InstanceFileInterface file = editInterface.getEditableProperties().First() as InstanceFileInterface;
+                file.Deleted = true;
+                String instanceFile = instanceFileManager.resolveSourceObject(editInterface);
+                InstanceWriter.Instance.deleteInstance(this, instanceFile);
+                removeInstanceFile(instanceFile);
+            }
         }
 
         private void duplicateSimObjectCallback(EditUICallback callback, EditInterfaceCommand command)
